Features/mapper xml (#52)
* Updates to new Xml mapper loader.

* Added support for loading from both XML and YML mappers.

* Updated TsGenerator to support arrays with an array node.

* WIP updates. Swapped things over to use MemoryAddress. Updated references to use uint64.

* WIP updates. Reworked Xml loading to be in stages.

* Updated integration tests. Moved some cleanup code to Startup out of the filesystem client notifier.

+++ src/GameHook.Domain/CommonEntities.cs
@@ -41,13 +41,13 @@ namespace GameHook.Domain
 
     public class GlossaryItem
     {
-        public GlossaryItem(uint key, object? value)
+        public GlossaryItem(ulong key, object? value)
         {
             Key = key;
             Value = value;
         }
 
-        public uint Key { get; private set; }
+        public ulong Key { get; private set; }
         public object? Value { get; private set; }
     }
 }
